// client/dist/widgets/arcgis/oriented-imagery-viewer/src/setting/constants.ts
import type { SettingInfo } from '../config'

export const settingsInfo: SettingInfo[] = [
  {
    name: 'viewerToolsEnabled',
    labelKey: 'viewerToolsToggleLabel',
    subTools: [
      {
        name: 'imageGalleryEnabled',
        labelKey: 'imageGalleryToolToggleLabel'
      },
      {
        name: 'imageEnahncementEnabled',
        labelKey: 'imageEnhancementToolToggleLabel'
      },
      {
        name: 'mapImageConversionToolEnabled',
        labelKey: 'mapImageConversionToolToggleLabel'
      },
      {
        name: 'showPopupsActionEnabled',
        labelKey: 'showPopupsActionToggleLabel'
      },
      {
        name: 'navigationToolEnabled',
        labelKey: 'navigationToolToggleLabel'
      },
      {
        name: 'directionalNavigationEnabled',
        labelKey: 'directionalNavigationToggleLabel'
      },
      {
        name: 'sequentialNavigationEnabled',
        labelKey: 'sequentialNavigationToggleLabel'
      },
      {
        name: 'measurementToolsEnabled',
        labelKey: 'measurementToolsToggleLabel'
      },
      {
        name: 'imageOverlaysEnabled',
        labelKey: 'imageOverlaysToggleLabel',
        subTools: [
          {
            name: 'dataCaptureEnabled',
            labelKey: 'digitizationToggleLabel'
          }
        ]
      }
    ]
  }
]
